requirements.txt
Flask==0.10.1
Frozen-Flask==0.11
Jinja2==2.7.1
MarkupSafe==0.18
Unidecode==0.04.16
Werkzeug==0.9.4
beautifulsoup4==4.3.2
boto==2.18.0
gdata==2.0.18
itsdangerous==0.23
python-slugify==0.0.7
requests==2.2.1
wsgiref==0.1.2
gspread==0.4.1
oauth2client==1.5.2
pycrypto==2.6.1
pyopenssl==16.0.0
